requirements.txt
dj-database-url==0.4.2
Django==1.11.2
djangorestframework==3.6.2
gunicorn==19.7.1
boto3==1.4.7
apscheduler==3.4.0
pendulum==1.4.1
psycopg2==2.7.4
sendgrid==5.6.0
requests==2.18.4
